Temporarily Disabled Loading error message for testing platform class loading (It is causing a lot of error messages and I haven't even done anything yet)
Set up variables to collect position in the LoadXmlItem in the Game class.

Set up proper x,y location loading in XmlLoad in item class.

Fixed infinite recursion in subclass XmlLoad functions.

Added a few things to Platform Class.

Fixed the issue of delayed closing of the game

Edit: Was messing around with the Add functions and forgot to put one back on last push. It caused a crash. Here is a fixed version of the last push.